SAT proofs: prune input early if necessary
When generating DRAT proofs from CaDiCaL, we only include
the clauses that are in the unsat core in the generated DIMACS
file.  Since those claues are alse the premisses of the corresponding
proof step, we need to prune them before generatign the proof step.

This adds an helper function to the SAT interface to check if
pruning is necessary, and prunes early if it is.